Fixes to comments
