+++ react/tic-tac-toe/src/app/screens/Game/components/Board/index.js
@@ -6,10 +6,13 @@ import styles from './styles.module.scss';
 
 class Board extends Component {
   renderSquare(i) {
+    const { squares, onClick } = this.props;
+
     return (
       <Square 
-        value={this.props.squares[i]} 
-        onClick={() => this.props.onClick(i)}
+        value={squares[i]} 
+        onClick={onClick}
+        id={i}
       />
     );
   }
